requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Machine Learning
scikit-learn==1.4.0
joblib==1.3.2

# Web Scraping
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.17.2
webdriver-manager==4.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
python-jose[cryptography]==3.3.0
